requirements.txt
matplotlib>=3.4.2
numpy>=1.21.0
Pillow>=8.3.1
pipreqs>=0.4.10
tensorflow>=2.5.0